Operators refactorisation
